+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGED:
 - Increased minimum PHP version requirement to 7.2
 - Yin 4.0.0 is minimally required
 - Update Zend Diactoros to 2.0
+- `JsonApiRequestValidatorMiddleware` will throw a `RequestException` if the request isn't instance of Yin's `RequestInterface`
 
 REMOVED: